+++ packages/cli/src/ui/App.tsx
@@ -84,7 +84,9 @@ import ansiEscapes from 'ansi-escapes';
 import { OverflowProvider } from './contexts/OverflowContext.js';
 import { ShowMoreLines } from './components/ShowMoreLines.js';
 import { PrivacyNotice } from './privacy/PrivacyNotice.js';
-import { updateEventEmitter } from '../utils/updateEventEmitter.js';
+import {
+  useUpdateHandler
+} from '../utils/handleAutoUpdate.js';
 
 const CTRL_EXIT_PROMPT_DURATION_MS = 1000;
 
@@ -110,69 +112,8 @@ const App = ({ config, settings, startupWarnings = [], version }: AppProps) => {
   const { history, addItem, clearItems, loadHistory } = useHistory();
 
   useEffect(() => {
-    let successfullyInstalled = false;
-    const handleUpdateRecieved = (info: UpdateInfo) => {
-      setUpdateInfo(info);
-      const savedMessage = info.message;
-      setTimeout((_: UpdateInfo) => {
-        if (!successfullyInstalled) {
-          addItem(
-            {
-              type: MessageType.INFO,
-              text: savedMessage,
-            },
-            Date.now(),
-          );
-        }
-        setUpdateInfo(null);
-      }, 60000);
-    };
-
-    const handleUpdateFailed = (_: UpdateInfo) => {
-      setUpdateInfo(null);
-      addItem(
-        {
-          type: MessageType.ERROR,
-          text: `Automatic update failed. Please try updating manually`,
-        },
-        Date.now(),
-      );
-    };
-
-    const handleUpdateSuccess = (_: UpdateInfo) => {
-      successfullyInstalled = true;
-      setUpdateInfo(null);
-      addItem(
-        {
-          type: MessageType.INFO,
-          text: `Update successful! The new version will be used on your next run.`,
-        },
-        Date.now(),
-      );
-    };
-
-    const handleUpdateInfo = (data: { message: string }) => {
-      addItem(
-        {
-          type: MessageType.INFO,
-          text: data.message,
-        },
-        Date.now(),
-      );
-    };
-
-    updateEventEmitter.on('update-received', handleUpdateRecieved);
-    updateEventEmitter.on('update-failed', handleUpdateFailed);
-    updateEventEmitter.on('update-success', handleUpdateSuccess);
-    updateEventEmitter.on('update-info', handleUpdateInfo);
-
-    return () => {
-      updateEventEmitter.off('update-received', handleUpdateRecieved);
-      updateEventEmitter.off('update-failed', handleUpdateFailed);
-      updateEventEmitter.off('update-success', handleUpdateSuccess);
-      updateEventEmitter.off('update-info', handleUpdateInfo);
-    };
-  }, [addItem]);
+    useUpdateHandler(addItem, setUpdateInfo);
+  }, [addItem])
 
   const {
     consoleMessages,
@@ -820,9 +761,7 @@ const App = ({ config, settings, startupWarnings = [], version }: AppProps) => {
 
         <Box flexDirection="column" ref={mainControlsRef}>
           {/* Move UpdateNotification to render update notification above input area */}
-          {updateInfo && (
-            <UpdateNotification message={`${updateInfo.message}`} />
-          )}
+          {updateInfo && <UpdateNotification message={updateInfo.message} />}
           {startupWarnings.length > 0 && (
             <Box
               borderStyle="round"
